+++ website/sidebars.js
@@ -1,0 +1,12 @@
+module.exports = {
+  someSidebar: {
+    ApiDef: [
+      "usage",
+      "config",
+      "middleware",
+      "caching",
+      "support",
+      "example",
+    ],
+  },
+};